arreglos finales

+++ rmdisk.cpp
@@ -25,9 +25,17 @@ void rmdisk::removedisk(){
     cin.get(i);
     cin.get(i);
     if(toupper(i)=='S'){
+        ifstream ejecutable;
+        ejecutable.open(ruta,ios::in);
+        if(ejecutable.fail()){
+            cout<<"El archivo no existe"<<endl;
+            return;
+        }
         string limpiar ="rm -fr \""+this->ruta+"\"";
         cout<<limpiar<<endl;
         system(limpiar.c_str());
+        
+        
         cout<<"Eliminacion ejecutada"<<endl;
     }else{
         cout<<"Eliminacion cancelada"<<endl;
